11dicembre

+++ Libft/ft_strnstr.c
@@ -6,7 +6,7 @@
 /*   By: stdi-pum <stdi-pum@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/28 18:19:15 by stdi-pum          #+#    #+#             */
-/*   Updated: 2023/11/28 19:44:05 by stdi-pum         ###   ########.fr       */
+/*   Updated: 2023/12/07 23:30:47 by stdi-pum         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 
 char	*ft_strnstr(const char *big, const char *little, size_t len)
 {
-	int		i;
+	size_t	i;
 	int		j;
 	char	*tbig;
 	char	*tlittle ;
